refactored my entity classes
