+++ src/shared/MessageType.java
@@ -3,4 +3,8 @@ package shared;
 public enum MessageType {
     ERROR_MESSAGE,
     SUCCESS_MESSAGE,
+    LIKE,
+    COMMENT,
+    NEW_INITIATIVE,
+    DIRECT_MESSAGE
 }
