--- src/shared/NotificationType.java
@@ -1,7 +1,0 @@
-package shared;
-
-public enum NotificationType {
-    LIKE,
-    COMMENT,
-    NEW_INITIATIVE
-}
